[feat](mips): refactor
