+++ include/llvm/asm/AsmPrinter.h
@@ -1,15 +1,12 @@
 ﻿#pragma once
 
-#include "asm/AsmWriter.h"
+#include "llvm/asm/AsmWriter.h"
 #include "llvm/ir/IrForward.h"
-#include "mips/mips_forward.h"
 #include <iostream>
 
 class AsmPrinter {
 public:
     void Print(ModulePtr module, std::ostream &out);
-    void Print(MipsData* data, std::ostream &out);
-    void Print(MipsCode* data, std::ostream &out);
 
 private:
     void _PrintHeader(AsmWriterPtr out);
