--- src/asm/MipsPrinter.cpp
@@ -1,4 +1,0 @@
-#include "mips/mips_inst.h"
-#include "mips/mips_reg.h"
-
-
