+++ src/llvm/asm/ValuePrinter.cpp
@@ -1,5 +1,5 @@
 ﻿#include "utils.h"
-#include "asm/AsmWriter.h"
+#include "llvm/asm/AsmWriter.h"
 #include "llvm/ir/Type.h"
 #include "llvm/ir/value/Argument.h"
 #include "llvm/ir/value/BasicBlock.h"
@@ -15,15 +15,15 @@
 
 #include <algorithm>
 
-void ValueRegTy::PrintAsm(AsmWriterPtr out) {
+void Value::PrintAsm(AsmWriterPtr out) {
     TOLANG_DIE("Operation not supported.");
 }
 
-void ValueRegTy::PrintName(AsmWriterPtr out) {
+void Value::PrintName(AsmWriterPtr out) {
     TOLANG_DIE("Operation not supported.");
 }
 
-void ValueRegTy::PrintUse(AsmWriterPtr out) {
+void Value::PrintUse(AsmWriterPtr out) {
     GetType()->PrintAsm(out);
     out->PushSpace();
     PrintName(out);
@@ -92,13 +92,13 @@ void Function::PrintAsm(AsmWriterPtr out) {
     out->Push('}').PushNewLine();
 }
 
-void ArgRegTy::PrintAsm(AsmWriterPtr out) {
+void Argument::PrintAsm(AsmWriterPtr out) {
     GetType()->PrintAsm(out);
     out->PushNext('%').Push(
         std::to_string(Parent()->GetSlotTracker()->Slot(this)));
 }
 
-void ArgRegTy::PrintUse(AsmWriterPtr out) { PrintAsm(out); }
+void Argument::PrintUse(AsmWriterPtr out) { PrintAsm(out); }
 
 void BasicBlock::PrintAsm(AsmWriterPtr out) {
     // We don't print the first basic block.
