+++ src/mips/MipsPrinter.cpp
@@ -1,0 +1,8 @@
+#include "mips/mips_inst.h"
+#include "mips/mips_reg.h"
+#include <iostream>
+
+void ZeroReg::PrintReg(std::ostream& out) {
+    out << "$zero";
+}
+
